tcti: Move private TCTI header from sysapi/include/ to tcti/.
This header should be private to the TCIT modules. The SAPI implementation
cannot rely on the structure of the private data in our TCTI contexts since
it must operate with TCTIs provided by others. The only thing that kept
the SAPI from exploding with TCTIs from other repos (i.e. the tpm2-abrmd) is
that, even though the SAPI was referencing the private TCTI structure, it was
only accessing function pointers from the common TCTI.

Moving this header to the tcti/ is intended to keep the other libraries from
using this data. All other changes in this commit are fixing up test cases
and inappropriate references to this file from the SAPI library.
