+++ test/unit/tcti-socket.c
@@ -33,7 +33,7 @@
 #include <cmocka.h>
 
 #include "sapi/tpm20.h"
-#include "sysapi/include/tcti_util.h"
+#include "tcti/tcti.h"
 #include "tcti/logging.h"
 #include "tcti/tcti_socket.h"
